backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
colorlog>=6.8.0
faiss-cpu>=1.7.4
numpy>=1.26.0
pydantic>=2.0.0
python-dotenv>=1.0.0
sse-starlette>=2.0.0
